Delete All Datasets

+++ api/model/storage/elastic/dataset.go
@@ -79,6 +79,7 @@ func (s *Storage) UpdateDataset(dataset *api.Dataset) error {
 		"clone":           dataset.Clone,
 		"immutable":       dataset.Immutable,
 		"parentDataset":   dataset.ParentDataset,
+		"deleted":         dataset.Deleted,
 	}
 
 	bytes, err := json.Marshal(source)
@@ -143,6 +144,7 @@ func (s *Storage) IngestDataset(datasetSource metadata.DatasetSource, meta *mode
 		"clone":            meta.Clone,
 		"immutable":        meta.Immutable,
 		"parentDataset":    "",
+		"deleted":          false,
 	}
 
 	bytes, err := json.Marshal(source)
@@ -164,10 +166,20 @@ func (s *Storage) IngestDataset(datasetSource metadata.DatasetSource, meta *mode
 }
 
 // DeleteDataset deletes a dataset from ES.
-func (s *Storage) DeleteDataset(dataset string) error {
-	_, err := s.client.Delete().Index(s.datasetIndex).Id(dataset).Refresh("true").Do(context.Background())
+func (s *Storage) DeleteDataset(dataset string, softDelete bool) error {
+	if !softDelete {
+		_, err := s.client.Delete().Index(s.datasetIndex).Id(dataset).Refresh("true").Do(context.Background())
+		return err
+	}
 
-	return err
+	// pull the dataset and update the deleted flag
+	ds, err := s.FetchDataset(dataset, true, true, true)
+	if err != nil {
+		return err
+	}
+	ds.Deleted = true
+
+	return s.UpdateDataset(ds)
 }
 
 func (s *Storage) parseDatasets(res *elastic.SearchResult, includeIndex bool, includeMeta bool, includeSystemData bool) ([]*api.Dataset, error) {
@@ -183,6 +195,16 @@ func (s *Storage) parseDatasets(res *elastic.SearchResult, includeIndex bool, in
 		if !ok {
 			id = hit.Id
 		}
+
+		// ignore deleted datasets
+		deleted, ok := json.Bool(src, "deleted")
+		if !ok {
+			deleted = false
+		}
+		if deleted {
+			continue
+		}
+
 		// extract the name
 		name, ok := json.String(src, "datasetName")
 		if !ok || name == "NULL" {
@@ -313,6 +335,7 @@ func (s *Storage) parseDatasets(res *elastic.SearchResult, includeIndex bool, in
 			Immutable:       immutable,
 			Clone:           clone,
 			ParentDataset:   parentDataset,
+			Deleted:         deleted,
 		})
 	}
 	return datasets, nil
